+++ src/components/LoginForm/LoginForm.jsx
@@ -1,30 +1,10 @@
 import React from "react";
 import { Form, Label, Input, Container } from "./LoginFormElements";
 import { Button } from "../Button/Button";
-import { useForm } from "react-hook-form";
-import * as yup from "yup";
-import { yupResolver } from "@hookform/resolvers/yup";
+import { useFields } from "./hooks/useFields";
 
 export const LoginForm = () => {
-  const schema = yup.object().shape({
-    email: yup
-      .string()
-      .email("Formato de email invalido")
-      .required("Campo requerido"),
-    password: yup.string().required("Campo requerido"),
-  });
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm({
-    resolver: yupResolver(schema),
-  });
-
-  const onSubmit = (values) => {
-    console.log("hola", values);
-  };
+  const { register, handleSubmit, onSubmit, errors } = useFields();
 
   return (
     <Container>
