+++ src/components/LoginForm/hooks/useFields.js
@@ -1,0 +1,30 @@
+import * as yup from "yup";
+import { useForm } from "react-hook-form";
+import { yupResolver } from "@hookform/resolvers/yup";
+import { useNavigate } from "react-router-dom";
+
+export const useFields = () => {
+  const navigate = useNavigate();
+
+  const schema = yup.object().shape({
+    email: yup
+      .string()
+      .email("Formato de email invalido")
+      .required("Campo requerido"),
+    password: yup.string().required("Campo requerido"),
+  });
+
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm({
+    resolver: yupResolver(schema),
+  });
+
+  const onSubmit = (values) => {
+    navigate("/");
+  };
+
+  return { register, handleSubmit, onSubmit, errors };
+};
