+++ src/components/Logo/Logo.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Images } from "../../assets/Images";
+import { Container } from "./LogoElements";
+
+export const Logo = () => {
+  return (
+    <Container>
+      <img src={Images.userIcon} alt="userIcon" className="userIcon" />
+      <h2>U control</h2>
+    </Container>
+  );
+};
